some light reorganization
