+++ tests/unit/CMakeLists.txt
@@ -16,16 +16,22 @@ target_compile_features(testdeps INTERFACE cxx_std_17)
 
 add_library(grtest_utils
   grtest_utils.hpp grtest_utils.cpp
+  grtest_os.hpp grtest_os.cpp
 )
 # we are being a little lazy with our usage of testdeps right here
 target_link_libraries(grtest_utils PUBLIC testdeps)
 target_compile_features(grtest_utils PUBLIC cxx_std_17)
 
-# short-term hack to get path input-file directory
-# (we can remove this hack once we introduce automatic file management in
-#  PR 235, PR 237, and PR 246)
+# these compile-definitions act as short-term hacks
 target_compile_definitions(grtest_utils
+  # this hack helps us get path input-file directory (we can remove it once we
+  # introduce automatic file management in PR 235, PR 237, and PR 246)
   PRIVATE GR_DATADIR=${CMAKE_CURRENT_SOURCE_DIR}/../../grackle_data_files/input/
+
+  # this hack lets us use Operating-system specific functionality (Once PR #237
+  # is merged, we should make use of the machinery introduced by that PR for
+  # enabling/disabling os-specific features)
+  PRIVATE "$<$<PLATFORM_ID:Linux,Darwin>:PLATFORM_GENERIC_UNIX>"
 )
 
 # start declaring targets for tests
@@ -36,12 +42,7 @@ target_link_libraries(runInterpolationTests testdeps)
 gtest_discover_tests(runInterpolationTests)
 
 add_executable(runStatusReporting test_status_reporting.cpp)
-target_link_libraries(runStatusReporting testdeps)
-# short-term hack. We should use the machinery introduced in PR #237 related to
-# Operating-system specific functionality when that PR is merged
-target_compile_definitions(runStatusReporting PRIVATE
-  "$<$<PLATFORM_ID:Linux,Darwin>:PLATFORM_GENERIC_UNIX>"
-)
+target_link_libraries(runStatusReporting grtest_utils)
 gtest_discover_tests(runStatusReporting)
 
 # one might argue that the following is more of an integration or end-to-end
